requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
et-xmlfile==1.1.0
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.5.0
google-cloud==0.34.0
google-cloud-bigquery==2.34.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gspread==5.2.0
httplib2==0.20.4
idna==3.3
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
proto-plus==1.20.3
protobuf==3.19.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tabulate==0.7.7
urllib3==1.21.1
xlrd==2.0.1
